Verifying jwt in App.js
I realized that I could change the jwt in localstorage and play with my name and id, adding verification has fixed this issue. If the jwt is not valid, on refresh I get logged out. This approach should be used in the api routes if you are thinking of creating a CRUD App.